+++ rebuild_PC/PsyCross/include/psx/libpress.h
@@ -6,10 +6,10 @@ extern "C" {
 #endif
 
 extern void DecDCTReset(int mode);
-extern int DecDCTBufSize(u_long *bs);
-extern void DecDCTvlc(u_long *bs, u_long *buf);
-extern void DecDCTin(u_long *buf, int mode);
-extern void DecDCTout(u_long *buf, int size);
+extern int DecDCTBufSize(u_int* bs);
+extern void DecDCTvlc(u_int* bs, u_int*buf);
+extern void DecDCTin(u_int* buf, int mode);
+extern void DecDCTout(u_int* buf, int size);
 extern int DecDCTinSync( int mode) ;
 extern int DecDCToutSync( int mode) ;
